requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
supabase
python-multipart
python-dotenv
pyjwt[crypto]